refactor: render message based on type inside system message omponent
